==> codes/0.3_map_microgenerators.Rmd <==
Map of all the micro-generators 


```{r}
rm(list=ls())
gc()
```

Upload the packages:
```{r}
pacman::p_load(sf, tidyverse, data.table, hrbrthemes, lwgeom, 
              maps, mapdata, spData, tigris, tidycensus, 
              leaflet, mapview, tmap, tmaptools, ggplot2, 
              geouy, readxl)
```


```{r}
theme_set(theme_bw())
```

Then, we load the directories:
```{r}
dir        <- list()
dir$root   <- dirname(getwd())
dir$data   <- paste0(dirname(getwd()), "/Data/")
dir$fig    <- paste0(dirname(getwd()), "/figures/")
dir$tab    <- paste0(dirname(getwd()), "/Tables/")
```

# Maps

Upload the Map of Uruguay
```{r}
uy = st_as_sf(
  map("world","uruguay", plot = FALSE, fill = TRUE))

uy_map = ggplot(uy) + 
  geom_sf(fill = "white", col = "blue", lwd = 0.3) +
labs(title = "The paisito" ,subtitle = "URUGUAY")

uy_map
```


# Microgenerators
```{r}
mg_solar <- sf::st_read(paste0(dir$data,
                               "Data_map_MG/generadores_micro_2025.kml"))

mg_solar <- 
  st_zm(mg_solar)

mg_solar = 
  st_transform(mg_solar, crs=st_crs(uy))

mg_solar = 
  st_crop(mg_solar, uy) # drop 2 observations which are outside the boundaries

map <- ggplot() + 
  geom_sf(data = uy, alpha = 0.8, 
          fill = "white", col = "black") + 
  geom_sf(data = mg_solar, col = "lightblue",  lwd = 1) + 
  labs(title = "Microgenerators in Uruguay")

map
```

```{r}
mg_solar_hhf <- mg_solar %>% 
  separate(Name, c("Name1", "Name2", "Name3", 
                   "Name4", "Name5", "Name6", "Name7"), 
           sep=" ", remove=FALSE)

mg_solar_hhf <- mg_solar_hhf %>% 
  mutate(firms_dummy =
                                        case_when(Name1=="FIDEICOMISO" ~1,
                                        Name1=="C.I.E.M.S.A." ~1,
                                        Name1=="FISCALIA" ~1,
                                        Name1=="U.T.E." ~ 1,
                                        Name1=="INTENDENCIA"~ 1,
                                        Name1=="COMISION" ~ 1, 
                                        Name1=="UTE" ~ 1,
                                        Name1=="CORPORACION"~ 1, 
                                        Name1=="CORP" ~ 1,    
                                        Name1=="CONSEJO"~ 1,       
                                        Name1=="TELEFONICA"~ 1,
                                        Name1=="FID"~ 1,
                                        Name1=="SA"~ 1,
                                        Name1=="ASOCIACION"~ 1,
                                        Name1=="COOPERATIVA"~ 1,
                                        Name1=="DUCSA"~ 1,
                                        Name1=="BBVA"~ 1,
                                        Name1=="SONARE,SRL"~ 1,
                                        Name1=="FACULTAD"~ 1,
                                        Name1=="ASOC.CIVIL"~ 1,
                                        Name1=="TELLAGORRYS.R.L."~ 1,
                                        Name1=="SOCIEDAD"~ 1,
                                        Name1=="CLUB"~ 1,
                                        Name1=="APOSTOLICA"~ 1,
                                        Name1=="S.A."~ 1,
                                        Name1=="SERVICIOS"~ 1,
                                        Name1=="HOPRESA" ~1,
                                        Name1=="NORTE" ~1,
                                        Name1=="AGROPECUARIA" ~1,
                                        Name1=="Telefónica"~1,
                                        Name1=="SUC.DE" ~1,
                                        Name1=="LOGIPARK,SA." ~1,
                                        Name1=="CENTRO" ~1,
                                        Name1=="CORP.DEL" ~1,
                                        Name1=="CONS.EDUC.TECNICO-PROFESIONAL" ~1,
                                        Name1=="ASPROE" ~1,
                                        Name1=="AVI-PORCLTDA" ~1,
                                        Name1=="MAOSOL.S.A."~1,
                                        Name1=="BLUMARY" ~1,
                                        Name1=="J.S.LIMTDA" ~1,
                                        Name1=="UTEC" ~1,
                                        Name1=="DISTRIBUIDORA" ~ 1,
                                        Name1=="DURNFORD," ~1,
                                        Name1=="HOGAR" ~1,
                                        Name1=="CANARIAS" ~1, 
                                        Name1=="COMISION" ~1, 
                                        Name1=="FILTROMAX.S.A.S" ~1, 
                                        Name1=="CERAMICAS" ~1, 
                                        Name1=="CADOL" ~1, 
                                        Name1=="Cooperativa" ~1, 
                                        Name1=="IJSUD" ~1, 
                                        Name1=="SOCOBIOMA" ~1, 
                                        Name1=="SOC.FOM.RURAL" ~1,
                                        Name1=="Cybe" ~1,
                                        Name1=="INSTITUT" ~1, 
                                        Name1=="PROLESA" ~1, 
                                        Name1=="AUTOMOVILES" ~1,
                                        Name1=="FUCAC" ~1,
                                        Name1=="FUNDACION" ~1, 
                                        Name1=="RADIO" ~1, 
                                        Name1=="AGRICOLA." ~1, 
                                    
                                        Name2=="S.A."~ 1,
                                        Name2=="S.A.,."~ 1,
                                        Name2=="S.A"~ 1,
                                        Name2=="SA"~ 1,
                                        Name2=="SA."~ 1,
                                        Name2=="LTDA."~ 1,
                                        Name2=="LTDA"~ 1,
                                        Name2=="S.R.L"~ 1, 
                                        Name2=="S.R.L."~ 1,
                                        Name2=="SRL"~ 1,  
                                        Name2=="LIMITADA"~ 1,   
                                        Name2=="FLORENCIA"~ 1,
                                        Name2=="SOCIEDAD"~ 1,
                                        Name2=="DESAFIO"~ 1,
                                        Name2=="CHACAREROS"~ 1,
                                        Name2=="PIAMONTES,SRL" ~1,
                                        Name2=="SAS"~ 1,
                                        Name2=="S.A," ~1, 
                                        Name2=="L.T.D.A" ~1, 
                                        Name2=="S.A.S"~ 1,
                                        Name2==",S.A" ~1, 
                                        Name2=="GENERAL" ~1, 
                                        Name2=="Sociedad" ~1,
                                        Name2=="S.A.S." ~1, 
                                        
                                               
                                        Name3=="S.A."~ 1,
                                        Name3=="S.A.,."~ 1,
                                        Name3=="S.A"~ 1,  
                                        Name3=="SA"~ 1,
                                        Name3=="SA."~ 1,       
                                        Name3=="LIMITADA"~ 1,
                                        Name3=="SOCIEDAD"~ 1,
                                        Name3=="COMERCIAL"~ 1,       
                                        Name3=="SOC."~ 1,
                                        Name3=="SOC"~ 1,
                                        Name3=="A"~ 1,       
                                        Name3=="LTDA"~ 1,       
                                        Name3=="LTDA."~ 1,
                                        Name3=="LTDA.,"~ 1,
                                        Name3=="S.R.L."~ 1,
                                        Name3=="S.R.L"~ 1,
                                        Name3=="SRL"~ 1,     
                                        Name3=="SCA"~ 1,
                                        Name3=="SCA,."~ 1,
                                        Name3=="ANONIMA"~ 1,
                                        Name3=="CIA"~ 1,
                                        Name3=="CIA.S.A"~ 1,
                                        Name3=="SAS" ~1,
                                        Name3=="S.A.S." ~1, 
                                        Name3=="IMM" ~1, 
                                        Name3=="S.A.S" ~1, 
                                               
                                        Name4=="S.A."~ 1,
                                        Name4=="S.A"~ 1,
                                        Name4=="S.A,"~ 1,
                                        Name4=="SA"~ 1, 
                                        Name4=="LTDA"~ 1,
                                        Name4=="LTDA."~ 1,  
                                        Name4=="ANONIMA"~ 1,   
                                        Name4=="S.R.L."~ 1,
                                        Name4=="S.R.L"~ 1,
                                        Name4=="SRL"~ 1,
                                        Name4=="SOCIEDAD"~ 1,   
                                        Name4=="CIA"~ 1,
                                        Name4=="SAS" ~1,
                                        Name4=="SR" ~1, 
                                        Name4=="LIMITADA" ~1, 
                                        Name4=="SAS" ~1,
                                               
                                        Name5=="NACION"~ 1,   
                                        Name5=="LTDA"~ 1,
                                        Name5=="LTDA."~ 1,       
                                        Name5=="S.A."~ 1,
                                        Name5=="S.A"~ 1,
                                        Name5=="SA"~ 1,
                                        Name5=="SRL"~ 1,       
                                        Name5=="S.R.L."~ 1,
                                        Name5=="SOCIEDAD" ~1, 
                                        
                                        Name6=="S.A."~ 1,
                                        Name6=="S.A"~ 1,
                                        Name6=="SA"~ 1,
                                        Name6=="SOCIEDAD"~ 1,
                                        Name6=="SUC"~ 1,
                                        Name6=="CLUB" ~1,
                                        Name6=="SRL" ~1, 
                                        
                                        Name7=="LIMITADA"~ 1 ,
                                        Name7=="S.A.," ~1, TRUE ~0))  

mg_solar_hhf <- mg_solar_hhf %>% 
  mutate(Agent =
           case_when(firms_dummy==1 ~"Firms",
                     TRUE~"Households"))
                                           
```

Check if the households are mostly households
```{r}
mg_solar_house <- 
  mg_solar_hhf %>% filter(firms_dummy==0) 

mg_solar_firm <- 
  mg_solar_hhf %>% filter(firms_dummy==1)
```


```{r}
all_uy <- load_geouy(c="Departamentos", 
                     crs = 32721, folder = tempdir())

all_uy = st_transform(all_uy, 
                      crs=st_crs(uy)) 

map2 <- ggplot() + 
  geom_sf(data = all_uy, alpha = 0.8, fill = "white", col = "black") +
  geom_sf(data = mg_solar_firm, lwd = 2 , aes(color = "Firms"), color = "skyblue") + 
  theme(text = element_text(size = 15),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks = element_blank()) +
  labs(title = "Solar Panels' Location")

map2
```

```{r}
ggsave(paste0(dir$fig,"map2.png"), 
       width = 7, height = 4.5)
```

Same map for Montevideo, the Capital City:
```{r}
mvdo <- all_uy %>% 
  filter(nombre=="MONTEVIDEO")

mg_solar_firm = 
  st_transform(mg_solar_firm, crs = st_crs(mvdo))

sf_use_s2(FALSE)
overlap <- 
  st_intersection(mvdo, mg_solar_firm)

map3 <- ggplot() +
  geom_sf(data = mvdo, alpha = 0.8, fill = "white", col = "black") +
  geom_sf(data = overlap, lwd = 2 , aes(color = Agent), color = "skyblue") +
  theme(text = element_text(size = 15),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks = element_blank()) +
  labs(title = "Solar Panels' Location",
       subtitle = "Montevideo")

map3
```

```{r}
ggsave(paste0(dir$fig, "map3.png"), 
       width = 7, height = 4.5)
```
 
## Add capacity

Arrange the capacity install by firms and hh:
```{r}
cap_data <- read_excel(paste0(dir$data,"cap_installed/cap_inst_data.xlsx"), 
                              sheet=2 ,range="A1:F2097")

cap_data <- cap_data %>% dplyr::rename("Name" = "Generador ",
                                "source" = "Fuente Primaria",
                                "state" = "Estado",
                                "power_aut" = "Potencia Autorizada",
                                "power_comp" = "Potencia Comprometida",
                                "power_inst" = "Potencia Instalada")

cap_data <- cap_data %>% separate(Name, c("Name1", "Name2", "Name3", "Name4", "Name5", "Name6", "Name7"), sep=" ", remove=FALSE)

cap_data <- cap_data %>% mutate(firms_dummy =
                                        case_when(Name1=="FIDEICOMISO" ~1,
                                        Name1=="C.I.E.M.S.A." ~1,
                                        Name1=="FISCALIA" ~1,
                                        Name1=="U.T.E." ~ 1,
                                        Name1=="INTENDENCIA"~ 1,
                                        Name1=="COMISION" ~ 1, 
                                        Name1=="UTE" ~ 1,
                                        Name1=="CORPORACION"~ 1, 
                                        Name1=="CORP" ~ 1,    
                                        Name1=="CONSEJO"~ 1,       
                                        Name1=="TELEFONICA"~ 1,
                                        Name1=="FID"~ 1,
                                        Name1=="SA"~ 1,
                                        Name1=="ASOCIACION"~ 1,
                                        Name1=="COOPERATIVA"~ 1,
                                        Name1=="DUCSA"~ 1,
                                        Name1=="BBVA"~ 1,
                                        Name1=="SONARE,SRL"~ 1,
                                        Name1=="FACULTAD"~ 1,
                                        Name1=="ASOC.CIVIL"~ 1,
                                        Name1=="TELLAGORRYS.R.L."~ 1,
                                        Name1=="SOCIEDAD"~ 1,
                                        Name1=="CLUB"~ 1,
                                        Name1=="APOSTOLICA"~ 1,
                                        Name1=="S.A."~ 1,
                                        Name1=="SERVICIOS"~ 1,
                                        Name1=="HOPRESA" ~1,
                                        Name1=="NORTE" ~1,
                                        Name1=="AGROPECUARIA" ~1,
                                        Name1=="Telefónica"~1,
                                        Name1=="SUC.DE" ~1,
                                        Name1=="LOGIPARK,SA." ~1,
                                        Name1=="CENTRO" ~1,
                                        Name1=="CORP.DEL" ~1,
                                        Name1=="CONS.EDUC.TECNICO-PROFESIONAL" ~1,
                                        Name1=="ASPROE" ~1,
                                        Name1=="AVI-PORCLTDA" ~1,
                                        Name1=="MAOSOL.S.A."~1,
                                        Name1=="BLUMARY" ~1,
                                        Name1=="J.S.LIMTDA" ~1,
                                        Name1=="UTEC" ~1,
                                        Name1=="DISTRIBUIDORA" ~ 1,
                                        Name1=="DURNFORD," ~1,
                                        Name1=="HOGAR" ~1,
                                        Name1=="CANARIAS" ~1, 
                                        Name1=="COMISION" ~1, 
                                        Name1=="FILTROMAX.S.A.S" ~1, 
                                        Name1=="CERAMICAS" ~1, 
                                        Name1=="CADOL" ~1, 
                                        Name1=="Cooperativa" ~1, 
                                        Name1=="IJSUD" ~1, 
                                        Name1=="SOCOBIOMA" ~1, 
                                        Name1=="SOC.FOM.RURAL" ~1,
                                        Name1=="Cybe" ~1,
                                        Name1=="INSTITUT" ~1, 
                                        Name1=="PROLESA" ~1, 
                                        Name1=="AUTOMOVILES" ~1,
                                        Name1=="FUCAC" ~1,
                                        Name1=="FUNDACION" ~1, 
                                        Name1=="RADIO" ~1, 
                                        Name1=="AGRICOLA." ~1, 
                                    
                                        Name2=="S.A."~ 1,
                                        Name2=="S.A.,."~ 1,
                                        Name2=="S.A"~ 1,
                                        Name2=="SA"~ 1,
                                        Name2=="SA."~ 1,
                                        Name2=="LTDA."~ 1,
                                        Name2=="LTDA"~ 1,
                                        Name2=="S.R.L"~ 1, 
                                        Name2=="S.R.L."~ 1,
                                        Name2=="SRL"~ 1,  
                                        Name2=="LIMITADA"~ 1,   
                                        Name2=="FLORENCIA"~ 1,
                                        Name2=="SOCIEDAD"~ 1,
                                        Name2=="DESAFIO"~ 1,
                                        Name2=="CHACAREROS"~ 1,
                                        Name2=="PIAMONTES,SRL" ~1,
                                        Name2=="SAS"~ 1,
                                        Name2=="S.A," ~1, 
                                        Name2=="L.T.D.A" ~1, 
                                        Name2=="S.A.S"~ 1,
                                        Name2==",S.A" ~1, 
                                        Name2=="GENERAL" ~1, 
                                        Name2=="Sociedad" ~1,
                                        Name2=="S.A.S." ~1, 
                                        
                                               
                                        Name3=="S.A."~ 1,
                                        Name3=="S.A.,."~ 1,
                                        Name3=="S.A"~ 1,  
                                        Name3=="SA"~ 1,
                                        Name3=="SA."~ 1,       
                                        Name3=="LIMITADA"~ 1,
                                        Name3=="SOCIEDAD"~ 1,
                                        Name3=="COMERCIAL"~ 1,       
                                        Name3=="SOC."~ 1,
                                        Name3=="SOC"~ 1,
                                        Name3=="A"~ 1,       
                                        Name3=="LTDA"~ 1,       
                                        Name3=="LTDA."~ 1,
                                        Name3=="LTDA.,"~ 1,
                                        Name3=="S.R.L."~ 1,
                                        Name3=="S.R.L"~ 1,
                                        Name3=="SRL"~ 1,     
                                        Name3=="SCA"~ 1,
                                        Name3=="SCA,."~ 1,
                                        Name3=="ANONIMA"~ 1,
                                        Name3=="CIA"~ 1,
                                        Name3=="CIA.S.A"~ 1,
                                        Name3=="SAS" ~1,
                                        Name3=="S.A.S." ~1, 
                                        Name3=="IMM" ~1, 
                                        Name3=="S.A.S" ~1, 
                                               
                                        Name4=="S.A."~ 1,
                                        Name4=="S.A"~ 1,
                                        Name4=="S.A,"~ 1,
                                        Name4=="SA"~ 1, 
                                        Name4=="LTDA"~ 1,
                                        Name4=="LTDA."~ 1,  
                                        Name4=="ANONIMA"~ 1,   
                                        Name4=="S.R.L."~ 1,
                                        Name4=="S.R.L"~ 1,
                                        Name4=="SRL"~ 1,
                                        Name4=="SOCIEDAD"~ 1,   
                                        Name4=="CIA"~ 1,
                                        Name4=="SAS" ~1,
                                        Name4=="SR" ~1, 
                                        Name4=="LIMITADA" ~1, 
                                        Name4=="SAS" ~1,
                                               
                                        Name5=="NACION"~ 1,   
                                        Name5=="LTDA"~ 1,
                                        Name5=="LTDA."~ 1,       
                                        Name5=="S.A."~ 1,
                                        Name5=="S.A"~ 1,
                                        Name5=="SA"~ 1,
                                        Name5=="SRL"~ 1,       
                                        Name5=="S.R.L."~ 1,
                                        Name5=="SOCIEDAD" ~1, 
                                        
                                        Name6=="S.A."~ 1,
                                        Name6=="S.A"~ 1,
                                        Name6=="SA"~ 1,
                                        Name6=="SOCIEDAD"~ 1,
                                        Name6=="SUC"~ 1,
                                        Name6=="CLUB" ~1,
                                        Name6=="SRL" ~1, 
                                        
                                        Name7=="LIMITADA"~ 1 ,
                                        Name7=="S.A.," ~1, TRUE ~0))  

cap_data <- cap_data %>% mutate(Agent =
                                    case_when(firms_dummy==1 ~"Firms", TRUE~"Households"))

cap_data_house <- cap_data %>% filter(firms_dummy==0) 
cap_data_firm <- cap_data %>% filter(firms_dummy==1)

cap_data_firm <- cap_data_firm %>% separate(power_inst, c("power_inst1", "power_inst_unit"), sep=" ", remove=FALSE)

cap_data_firm$power_inst1 <- as.numeric(cap_data_firm$power_inst1)

cap_data_firm <- cap_data_firm %>% dplyr::select(-c("power_comp")) 

cap_data_firm <- cap_data_firm %>% mutate(Name=case_when(
 Name=="FIERRO VIGNOLI S.A. 1" ~ "FIERRO VIGNOLI S.A.",   TRUE ~ Name))

cap_data_firm <- cap_data_firm %>% filter(Name != "ALWAYS HOTEL LTDA.",
                                          Name != "ARAHAKU S.R.L", 
                                          Name != "CRISTALDEK S.A.",
                                          Name != "DIVINO S.A.",
                                          Name != "FIERRO VIGNOLI S.A.",
                                          Name != "NOMAGU SA",
                                          Name != "OKEIDELIYCAFE SRL",
                                          Name != "PASEO DEL PUERTO S.A.")

mg_solar_firm <- 
  mg_solar_firm %>% 
  filter(Name != "FIERRO VIGNOLI S.A.")
```

Merge both datasets: 
```{r}
mg_solar_firm <- 
  mg_solar_firm %>% arrange(Name) #sort
mg_solar_firm <- 
  mg_solar_firm %>% mutate(ID = row_number()) #add ID number
mg_solar_firm <- 
  mg_solar_firm %>% select(Name, ID, everything()) # order the columns

cap_data_firm <- cap_data_firm %>% 
  arrange(Name)
cap_data_firm <- cap_data_firm %>% 
  mutate(ID = row_number())
cap_data_firm <- cap_data_firm %>% 
  select(Name, ID, everything())

cap_data_firm <- cap_data_firm %>% 
  dplyr::select(-c("Name1", "Name2", "Name3", "Name4", "Name5", "Name6", "Name7", "power_inst")) 

mg_solar_firm <- mg_solar_firm %>% 
  dplyr::select(-c("Name1", "Name2", "Name3", "Name4", "Name5", "Name6", "Name7", "Description")) 

cap_data_merge <- 
  base::merge(mg_solar_firm, cap_data_firm,  by="ID")
```

 
 
```{r}
cap_data_merge <- cap_data_merge %>% 
  mutate(power_inst2=power_inst1*1000) # to go from MW to Kw

cap_data_merge <- cap_data_merge %>% 
  rename("Power" = "power_inst2" )

map4 <- ggplot() + 
  geom_sf(data = all_uy, alpha = 0.8, fill = "white", col = "black") +
  geom_sf(data = cap_data_merge, aes(color = "Firms", size = Power), color = "skyblue") + 
  theme(text = element_text(size = 15),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks = element_blank()) +
  labs(title = "Solar Panels' Location")

map4
```


```{r}
ggsave(paste0(dir$fig, "map4.png"), 
       width = 7, height = 4.5)
```
 
```{r}
mvdo <- 
  all_uy %>% 
  filter(nombre=="MONTEVIDEO")

cap_data_merge = 
  st_transform(cap_data_merge, crs = st_crs(mvdo))

sf_use_s2(FALSE)
overlap2 <- 
  st_intersection(mvdo, cap_data_merge)

map5 <- ggplot() +
  geom_sf(data = mvdo, alpha = 0.8, fill = "white", col = "black") +
  geom_sf(data = overlap2, aes(color = Agent, size = Power), color="skyblue") +
  theme(text = element_text(size = 15),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks = element_blank()) +
  labs(title = "Solar Panels' Location",
       subtitle = "Montevideo")

map5
```


```{r}
ggsave(paste0(dir$fig, "map5.png"), 
       width = 7, height = 4.5)
```

## Firms
```{r}

cap_data_firm <- cap_data_firm %>% mutate(Power=power_inst1*1000)

cap_data_firm %>% 
  summarize(mean_ext=mean(cap_data_firm$Power, na.rm=TRUE), 
            sd_ext=sd(cap_data_firm$Power, na.rm=TRUE), 
            min_ext=min(cap_data_firm$Power, na.rm=TRUE), 
            max_ext=max(cap_data_firm$Power, na.rm=TRUE))
```

Total production
```{r}
cap_data_firm %>% 
  summarize(mean_ext=mean(cap_data_firm$power_inst1, na.rm=TRUE),
            sd_ext=sd(cap_data_firm$power_inst1, na.rm=TRUE),
            min_ext=min(cap_data_firm$power_inst1, na.rm=TRUE),
            max_ext=max(cap_data_firm$power_inst1, na.rm=TRUE))
```
